Self-Confidence for candidates
Now each candidate has a self-confidence level which will be increased on every positive feedback (as well as a completely distributed campaign) and decreased on every negative feedback. The candidate decides whether to start a campaign or to send a choose-me-message depending on his self-confidence level. Is the level positive (or 0) he sends a choose-me-message otherwise a campaign. After sending a campaign/message the level will be reset to 0.